Refactor : step 3에서 리팩토링으로 변경
