+++ src/main/java/step3/Names.java
@@ -1,4 +1,4 @@
-package step4;
+package step3;
 
 import java.util.ArrayList;
 import java.util.Arrays;
