+++ src/main/java/step3/UserInput.java
@@ -6,29 +6,30 @@ public class UserInput {
     private final static String NUMBER_PATTERN = "\\d*$";
     private final static String NUMBER_EXCEPTION_MESSAGE = "입력 값이 숫자가 아닙니다.";
     private final static String NULL_EXCEPTION_MESSAGE = "입력 값이 없습니다.";
+    private final static String NAMES_NOT_HAVE_COMMA_MESSAGE = "이름 값은 ,로 구분됩니다.";
 
-    private final String countStr;
+    private final String names;
     private final String timeStr;
 
-    private UserInput(String countStr, String timeStr) {
-        this.countStr = countStr;
+    private UserInput(String names, String timeStr) {
+        this.names = names;
         this.timeStr = timeStr;
     }
 
-    public static UserInput create(String countStr, String timeStr) {
-        return new UserInput(countStr, timeStr);
+    public static UserInput create(String names, String timeStr) {
+        return new UserInput(names, timeStr);
     }
 
     public void validate() {
-        isNullOrEmptyForString(countStr);
-        isNumberForString(countStr);
+        isNullOrEmptyForString(names);
+        isExistComma(names);
 
         isNullOrEmptyForString(timeStr);
         isNumberForString(timeStr);
     }
 
-    public String getCountStr() {
-        return countStr;
+    public String getNames() {
+        return names;
     }
 
     public String getTimeStr() {
@@ -47,9 +48,9 @@ public class UserInput {
         }
     }
 
-    private void isNull(Integer input) {
-        if (input == null) {
-            throw new IllegalArgumentException(NULL_EXCEPTION_MESSAGE);
+    private void isExistComma(String input) {
+        if (!input.contains(",")) {
+            throw new IllegalArgumentException(NAMES_NOT_HAVE_COMMA_MESSAGE);
         }
     }
 }
